+++ CHANGELOG.md
@@ -7,13 +7,12 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 
 **New fixes in 1.25.x:**
 
-* [scifi5097: Username is hard to read in comments posted by the OP on meta](http://meta.scifi.stackexchange.com/q/5097) (CSS, meta.scifi.SE only)
 * [salesforce836: Div containing 2 minute tour button not big enough](http://meta.salesforce.stackexchange.com/q/836) (CSS, salesforce.SE only)
 * [mse240417: Inside or outside?](http://meta.stackoverflow.com/q/240417)
 * [mse240102: Lightbox sometimes doesn't go away when popup is closed in user profile editor](http://meta.stackoverflow.com/q/240102)
-* [mse84296: RTL text can mess up comment timestamps](http://meta.stackexchange.com/q/84296) (CSS)
 * [mse240710: Was the fringe always there on the up-rep icon?](http://meta.stackexchange.com/q/240710) (CSS)
 * [mse240790: Welcome back, user. Click here to get error 404](http://meta.stackoverflow.com/q/240790)
+* [mse243519: Dangling signature dash in comments](http://meta.stackoverflow.com/q/243519)
 
 **Other changes:**